+++ CHANGELOG.md
@@ -13,6 +13,7 @@ modified Semantic Versioning scheme. See the "Versioning scheme" section of the
 - RIG-6: Added shivammathur/setup-php PHP setup action to develop workflow to lock php at 7.3.
 
 ### Changed
+- RIG-130: Changed notifications block to use node langcode rather than the current users.
 
 ### Deprecated